fixed int to float error in scale

+++ inc/april_tags.hpp
@@ -31,7 +31,7 @@ struct Point2D
 class AprilTagDetector 
 {
 public:
-    AprilTagDetector(int tag_size_corners, int tag_size_center, int offset) 
+    AprilTagDetector(float tag_size_corners, float tag_size_center, float offset) 
     : tag_size_corners(tag_size_corners), tag_size_center(tag_size_center), offset(offset)
     {
        // tag_offset = offset / (1.0 * tag_size_corners);
@@ -70,9 +70,9 @@ public:
 private:
     apriltag_family_t *tf;
     apriltag_detector_t *td;
-    int tag_size_corners; // in cm
-    int tag_size_center; // in cm
-    int offset; // in terms of tag_size_corners
+    float tag_size_corners; // in cm
+    float tag_size_center; // in cm
+    float offset; // in terms of tag_size_corners
     float tag_offset;
     std::unordered_map<int, Point2D> tag_positions;
     std::unordered_map<int, float> tag_sizes;
